# Faster format for loading code