--- a/c/Mf-base
+++ b/c/Mf-base
@@ -23,7 +23,7 @@ Main=../boot/$m/main.$o
 Scheme=../bin/$m/scheme
 
 kernelsrc=statics.c segment.c alloc.c symbol.c intern.c gcwrapper.c gc-ocd.c gc-oce.c\
- number.c schsig.c io.c new-io.c print.c fasl.c stats.c foreign.c prim.c prim5.c flushcache.c\
+ number.c schsig.c io.c new-io.c print.c fasl.c vfasl.c stats.c foreign.c prim.c prim5.c flushcache.c\
  schlib.c thread.c expeditor.c scheme.c
 
 kernelobj=${kernelsrc:%.c=%.$o} ${mdobj}