… c/vfasl.c