The good news is that this patch can cut startup time in half when loading uncompressed boot files. The bad news is that it's a significant addition to the kernel: about 2/3 the size of the GC implementation, and duplicating some of the GC code in its current form.

On my machine, startup time from uncompressed "petite.boot" and "scheme.boot" files is reduced from 60ms to 30ms. With Racket-on-Chez's "racket.boot" in addition (which is a little larger than "petite.boot" and "scheme.boot" combined), startup goes from 140ms to 70ms. The extra code seems worth it to save 70ms on every Racket's startup — unless, of course, something simpler can provide the same benefit.

The "vfasl" format implemented by this patch is mostly an image of the code and data that can read directly into a GC-managed memory space. Just as important as loading the image directly, it can be loaded into the static generation, saving work to compact loaded code into the static generation.
(The vfasl loader does not update object counts for static content.)

The vfasl format can also cut loading non-boot code in half, at least for relatively large code files, but the benefit of vfasl loading is limited if the code has to be traversed later by the GC. Also, the in-memory format of data is larger than the fasl format, and the size difference tends to defeat any benefit of vfasl for small code files.

Use `vfasl-convert-file` to convert to the vfasl format, something like this:
 ```
     (compile-compressed #f)
     (vfasl-convert-file "orig/petite.boot" "new/petite.boot" '())
     (vfasl-convert-file "orig/scheme.boot" "new/scheme.boot" '("petite"))
     (vfasl-convert-file "orig/mine.boot" "new/mine.boot" '("petite" "scheme"))
```