@samth 2018-12-31
This looks potentially related to work in GHC: http://ezyang.com/compact.html I wonder if they managed to avoid the complexity you're worried about.