@mflatt 2018-12-31
@samth That work doesn't deal with code, shared objects like symbols and record-type descriptors, or mutable values (such as symbols that have property lists) – all ways to reduce complexity, certainly, but not applicable to the problem of loading and linking Scheme code.