@dybvig 2019-01-18
For many years, Chez Scheme supported the notion of "saved heaps", and "incremental saved heaps", which is similar to this.  Saved heaps were essentially pre-built heap images.  There was a saved heap that corresponded to the base boot file and incremental saved heaps that corresponded to non-base boot files.  With mmap'd saved heaps, startup time was practically zero, and sharing of memory between processes cut total memory usage considerably when many Scheme processes were used.

We ended up abandoning saved heaps when address randomization became ubiquitous.

Perhaps it's time to come up with a new implementation of saved heaps.  One approach is to have a "live" saved heap in the form of a background process created upon the first invocation of Chez Scheme (or Racket), and have each subsequent invocations merely fork a new copy of the background process.  This has the downside that the forked process is a child of the background process rather than the program (e.g., shell) of invocation and its environment, but it's not unlike how other applications work these days.

Another approach is to save an image as we used to and adjust all the pointers on start-up to account for the new random base address.  Of course, we can always request the original address when mmap'ing and avoid the adjustment if we get it.

If machines supported EA-relative addressing (like PC-relative jumps add in the address of the jump, EA-relative loads would add the effective address of a loaded pointer to the loaded pointer), we could avoid the adjustment entirely.  We could of course do the add ourselves at the cost of an extra instruction or two at each ptr load.