@dybvig 2019-02-02
Thanks for the suggestion, @mromyers.  I'll take a look at the slides when I get a chance, though I know how to go about saving the Chez Scheme heap from prior experience.

A warning for anyone else considering working on saved heaps for Chez Scheme: The slides are as deep as we can go, unless there's a more detailed writeup.  Unless I'm mistaken, the Emacs source code is GPL'd, so even if someone just perused the code to gain knowledge of its workings before working on an implementation for Chez Scheme, that could taint our (permissive open-source) code.