@fweimer 2020-11-16
Perhaps the comment at the top of `c/fasl.c` should be updated to mention the vfasl alternative?

I've started to look at translating existing fasl files to `ET_REL` ELF objects, basically letting the ELF link editor handle the relocation processing. I do not have any code yet, but I believe that ELF is sufficiently powerful for this (e.g., large section alignment to meet GC requirements). The initial goal would be to produce position-dependent code only, which simplifies the task somewhat.