@dybvig 2020-11-18
I'll be interested in seeing where you end up with this.  Code objects presently contain pointers to other objects, e.g., quoted constants, symbols used to store top-level values, and other code objects.  It will be difficult to support all of those with position-independent code without going to relocation tables.

Whether ELF is powerful enough or not, we might want to go with a solution of our own for portability.