@LiberalArtist 2021-04-13
Just as another point in the design space, according to [Andy Wingo's blog](https://wingolog.org/archives/2014/01/19/elf-in-guile), Guile has been using ELF as its object file format for several years. For portability, apparently they implement their own linker and loader and use them on all platforms.