@mflatt 2021-05-05
While vfasl has worked well for Racket, the implementation in this PR is no longer the one used. The main change is that the vfasl writer was rewritten in Scheme. (See racket/ChezScheme@729dbd09ba35f48c45ab1c060a62178a68a6ad16 for more information.) So, I'll close this PR.